+++ application/src/main/java/com/course/admin/catalogo/application/category/retrieve/get/GetCategoryByIdUseCase.java
@@ -1,7 +1,6 @@
 package com.course.admin.catalogo.application.category.retrieve.get;
 
 import com.course.admin.catalogo.application.UseCase;
-import com.course.admin.catalogo.application.category.retrieve.CategoryOutput;
 
 public abstract class GetCategoryByIdUseCase extends UseCase<String, CategoryOutput> {
 }
